Adapt an iterator with transformation steps

iter_adapt() takes an iterator iter and a list of Transformation steps. It returns an iterator that returns transformed values.

iter_adapt(iter, ...)

Arguments

iter
An iterator.

...
Transformation steps. These dots are taken with implicit splicing of lists and passed to compose().

See also

Examples

# Let's create a simple iterator:
iter <- as_iterator(1:50)
iter()

#> [1] 1

# We can transform it by mapping functions over the elements:
iter <- iter_adapt(iter, iter_map(`+`, 100L))
take(iter, 5)

#> [[1]]
#> [1] 102
#> 
#> [[2]]
#> [1] 103
#> 
#> [[3]]
#> [1] 104
#> 
#> [[4]]
#> [1] 105
#> 
#> [[5]]
#> [1] 106
#> 

# Or by discarding unwanted values:
iter <- iter_adapt(iter, iter_discard(~ .x %% 2 == 0))
take(iter, 5)

#> [[1]]
#> [1] 107
#> 
#> [[2]]
#> [1] 109
#> 
#> [[3]]
#> [1] 111
#> 
#> [[4]]
#> [1] 113
#> 
#> [[5]]
#> [1] 115
#> 

# Note that iter_adapt() accepts several transformation at once and
# can of course adapt generators which are regular iterators:
iter <- gen(for (x in 1:10) yield(x))
iter <- iter_adapt(iter,
  iter_map(`+`, 100L),
  iter_discard(~ .x %% 2 == 0)
)
take(iter, 5)

#> [[1]]
#> [1] 101
#> 
#> [[2]]
#> [1] 103
#> 
#> [[3]]
#> [1] 105
#> 
#> [[4]]
#> [1] 107
#> 
#> [[5]]
#> [1] 109
#>