Transformation steps

Transformation steps can be chained together to modify the behaviour of an iterator (see iter_adapt()).

iter_map() applies a function .f over all inputs.

iter_discard() and iter_keep() apply a predicate .p over inputs and discard or keep the selected elements.

iter_take() is a transformation step that terminates early after n inputs.

iter_take(n)

iter_map(.f, ...)

iter_keep(.p, ...)

iter_discard(.p, ...)

Arguments

n
The number of inputs to take.

.f
A function to map over inputs. If needed this argument is transformed to a function with rlang::as_closure() and thus supports the lambda-formula notation.

...
Further arguments passed over to .f or .p.

.p
A predicate function applied to inputs. If needed this argument is transformed to a function with rlang::as_closure() and thus supports the lambda-formula notation.

Details

Transformation steps, also called transducers, are function operators that take a reducer function and return another reducer function with modified behaviour.

Examples

# `purrr::compose()` is the recommended way to chain transformation
# steps:
steps <- purrr::compose(iter_map(`+`, 10), iter_discard(`>`, 15))