Take n or all elements from a vector or iterator

These functions are similar to head() but also perform coercion to a given output type. They support all reducible objects, including iterators. In fact their main purpose is to take or drain elements of an iterator or generator into a proper vector.

The take() variants return a defined number of elements.

The drain() variants return all elements. The vector is grown geometrically.

take(x, n)

take_lgl(x, n)

take_int(x, n)

take_dbl(x, n)

take_cpl(x, n)

take_chr(x, n)

take_raw(x, n)

drain(x)

drain_lgl(x)

drain_int(x)

drain_dbl(x)

drain_cpl(x)

drain_chr(x)

drain_raw(x, n)

Arguments

x
A reducible object.

n
The number of elements to take from x.

See also

Examples

take(letters, 5)

#> [[1]]
#> [1] "a"
#> 
#> [[2]]
#> [1] "b"
#> 
#> [[3]]
#> [1] "c"
#> 
#> [[4]]
#> [1] "d"
#> 
#> [[5]]
#> [1] "e"
#> 

take_chr(1:10, 5)

#> [1] "1" "2" "3" "4" "5"

iter <- as_iterator(1:10)
drain(iter)

#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] 3
#> 
#> [[4]]
#> [1] 4
#> 
#> [[5]]
#> [1] 5
#> 
#> [[6]]
#> [1] 6
#> 
#> [[7]]
#> [1] 7
#> 
#> [[8]]
#> [1] 8
#> 
#> [[9]]
#> [1] 9
#> 
#> [[10]]
#> [1] 10
#>